+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # nextjs
 
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies [3595d0f]
+  - @godaddy/react@1.0.47
+
 ## 0.1.44
 
 ### Patch Changes